feat: separate client components from validate function (use client issue)
