+++ internal/internal/src/index.ts
@@ -2,3 +2,8 @@ import "./styles.css";
 
 export * as Metadata from "./metadata";
 export { ContactUsForm } from "./contact-form";
+
+export {
+  validateContactFormData,
+  type ContactFormDataProps,
+} from "./contact-form-validation";
